Remove duplication in wiki API tests
