+++ api_tests/nodes/views/test_node_wiki_detail.py
@@ -1,20 +1,12 @@
 from nose.tools import *  # flake8: noqa
 
 from api.base.settings.defaults import API_BASE
-from tests.base import ApiTestCase
-from tests.factories import (AuthUserFactory, ProjectFactory, RegistrationFactory,
-                             NodeWikiFactory, PrivateLinkFactory)
+from api_tests.wikis.views.test_wiki_detail import TestWikiDetailMixin
+from tests.base import ApiWikiTestCase
+from tests.factories import ProjectFactory, RegistrationFactory, NodeWikiFactory
 
 
-class TestNodeWikiDetailView(ApiTestCase):
-    def setUp(self):
-        super(TestNodeWikiDetailView, self).setUp()
-        self.user = AuthUserFactory()
-        self.non_contributor = AuthUserFactory()
-
-    def _add_project_wiki_page(self, node, user):
-        # API will only return current wiki pages
-        return NodeWikiFactory(node=node, user=user, is_current=True)
+class TestNodeWikiDetailView(ApiWikiTestCase, TestWikiDetailMixin):
 
     def _set_up_public_project_with_wiki_page(self):
         self.public_project = ProjectFactory(is_public=True, creator=self.user)
@@ -41,98 +33,6 @@ class TestNodeWikiDetailView(ApiTestCase):
         self.private_registration.wiki_pages_current = {'home': self.private_registration_wiki_id}
         self.private_registration.save()
         self.private_registration_url = '/{}nodes/{}/wikis/{}/'.format(API_BASE, self.private_registration._id, self.private_registration_wiki_id)
-
-    def test_public_node_logged_out_user_can_view_wiki(self):
-        self._set_up_public_project_with_wiki_page()
-        res = self.app.get(self.public_url)
-        assert_equal(res.status_code, 200)
-        assert_equal(res.json['data']['id'], self.public_wiki._id)
-
-    def test_public_node_logged_in_non_contributor_can_view_wiki(self):
-        self._set_up_public_project_with_wiki_page()
-        res = self.app.get(self.public_url, auth=self.non_contributor.auth)
-        assert_equal(res.status_code, 200)
-        assert_equal(res.json['data']['id'], self.public_wiki._id)
-
-    def test_public_node_logged_in_contributor_can_view_wiki(self):
-        self._set_up_public_project_with_wiki_page()
-        res = self.app.get(self.public_url, auth=self.user.auth)
-        assert_equal(res.status_code, 200)
-        assert_equal(res.json['data']['id'], self.public_wiki._id)
-
-    def test_private_node_logged_out_user_cannot_view_wiki(self):
-        self._set_up_private_project_with_wiki_page()
-        res = self.app.get(self.private_url, expect_errors=True)
-        assert_equal(res.status_code, 401)
-        assert_equal(res.json['errors'][0]['detail'], 'Authentication credentials were not provided.')
-
-    def test_private_node_logged_in_non_contributor_cannot_view_wiki(self):
-        self._set_up_private_project_with_wiki_page()
-        res = self.app.get(self.private_url, auth=self.non_contributor.auth, expect_errors=True)
-        assert_equal(res.status_code, 403)
-        assert_equal(res.json['errors'][0]['detail'], 'You do not have permission to perform this action.')
-
-    def test_private_node_logged_in_contributor_can_view_wiki(self):
-        self._set_up_private_project_with_wiki_page()
-        res = self.app.get(self.private_url, auth=self.user.auth)
-        assert_equal(res.status_code, 200)
-        assert_equal(res.json['data']['id'], self.private_wiki._id)
-
-    def test_private_node_user_with_anonymous_link_can_view_wiki(self):
-        self._set_up_private_project_with_wiki_page()
-        private_link = PrivateLinkFactory(anonymous=True)
-        private_link.nodes.append(self.private_project)
-        private_link.save()
-        res = self.app.get('/{}nodes/{}/wikis/{}/'.format(API_BASE, self.private_project._id, self.private_wiki._id),
-                           {'view_only': private_link.key})
-        assert_equal(res.status_code, 200)
-        assert_equal(res.json['data']['id'], self.private_wiki._id)
-
-    def test_private_node_user_with_view_only_link_can_view_wiki(self):
-        self._set_up_private_project_with_wiki_page()
-        private_link = PrivateLinkFactory(anonymous=False)
-        private_link.nodes.append(self.private_project)
-        private_link.save()
-        res = self.app.get('/{}nodes/{}/wikis/{}/'.format(API_BASE, self.private_project._id, self.private_wiki._id),
-                           {'view_only': private_link.key})
-        assert_equal(res.status_code, 200)
-        assert_equal(res.json['data']['id'], self.private_wiki._id)
-
-    def test_public_registration_logged_out_user_cannot_view_wiki(self):
-        self._set_up_public_registration_with_wiki_page()
-        res = self.app.get(self.public_registration_url, expect_errors=True)
-        assert_equal(res.status_code, 200)
-        assert_equal(res.json['data']['id'], self.public_registration_wiki_id)
-
-    def test_public_registration_logged_in_non_contributor_cannot_view_wiki(self):
-        self._set_up_public_registration_with_wiki_page()
-        res = self.app.get(self.public_registration_url, auth=self.non_contributor.auth, expect_errors=True)
-        assert_equal(res.status_code, 200)
-        assert_equal(res.json['data']['id'], self.public_registration_wiki_id)
-
-    def test_public_registration_contributor_can_view_wiki(self):
-        self._set_up_public_registration_with_wiki_page()
-        res = self.app.get(self.public_registration_url, auth=self.user.auth)
-        assert_equal(res.status_code, 200)
-        assert_equal(res.json['data']['id'], self.public_registration_wiki_id)
-
-    def test_private_registration_logged_out_user_cannot_view_wiki(self):
-        self._set_up_private_registration_with_wiki_page()
-        res = self.app.get(self.private_registration_url, expect_errors=True)
-        assert_equal(res.status_code, 401)
-        assert_equal(res.json['errors'][0]['detail'], 'Authentication credentials were not provided.')
-
-    def test_private_registration_logged_in_non_contributor_cannot_view_wiki(self):
-        self._set_up_private_registration_with_wiki_page()
-        res = self.app.get(self.private_registration_url, auth=self.non_contributor.auth, expect_errors=True)
-        assert_equal(res.status_code, 403)
-        assert_equal(res.json['errors'][0]['detail'], 'You do not have permission to perform this action.')
-
-    def test_private_registration_contributor_can_view_wiki(self):
-        self._set_up_private_registration_with_wiki_page()
-        res = self.app.get(self.private_registration_url, auth=self.user.auth)
-        assert_equal(res.status_code, 200)
-        assert_equal(res.json['data']['id'], self.private_registration_wiki_id)
 
     def test_wiki_invalid_id_not_found(self):
         url = '/{}nodes/{}wikis/{}/'.format(API_BASE, ProjectFactory()._id, 'abcde')
