+++ api_tests/wikis/views/test_wiki_detail.py
@@ -1,47 +1,25 @@
+import furl
 from urlparse import urlparse
 from nose.tools import *  # flake8: noqa
 
 from api.base.settings.defaults import API_BASE
-from tests.base import ApiTestCase
-from tests.factories import (AuthUserFactory, ProjectFactory, RegistrationFactory,
+from tests.base import ApiWikiTestCase
+from tests.factories import (ProjectFactory, RegistrationFactory,
                              NodeWikiFactory, PrivateLinkFactory)
 
 
-class TestWikiDetailView(ApiTestCase):
-    def setUp(self):
-        super(TestWikiDetailView, self).setUp()
-        self.user = AuthUserFactory()
-        self.non_contributor = AuthUserFactory()
-
-    def _add_project_wiki_page(self, node, user):
-        # API will only return current wiki pages
-        return NodeWikiFactory(node=node, user=user, is_current=True)
-
+class TestWikiDetailMixin(object):
     def _set_up_public_project_with_wiki_page(self):
-        self.public_project = ProjectFactory(is_public=True, creator=self.user)
-        self.public_wiki = self._add_project_wiki_page(self.public_project, self.user)
-        self.public_url = '/{}wikis/{}/'.format(API_BASE, self.public_wiki._id)
+        raise NotImplementedError
 
     def _set_up_private_project_with_wiki_page(self):
-        self.private_project = ProjectFactory(creator=self.user)
-        self.private_wiki = self._add_project_wiki_page(self.private_project, self.user)
-        self.private_url = '/{}wikis/{}/'.format(API_BASE, self.private_wiki._id)
+        raise NotImplementedError
 
     def _set_up_public_registration_with_wiki_page(self):
-        self._set_up_public_project_with_wiki_page()
-        self.public_registration = RegistrationFactory(project=self.public_project, user=self.user, is_public=True)
-        self.public_registration_wiki_id = self.public_registration.wiki_pages_versions['home'][0]
-        self.public_registration.wiki_pages_current = {'home': self.public_registration_wiki_id}
-        self.public_registration.save()
-        self.public_registration_url = '/{}wikis/{}/'.format(API_BASE, self.public_registration_wiki_id)
+        raise NotImplementedError
 
     def _set_up_private_registration_with_wiki_page(self):
-        self._set_up_private_project_with_wiki_page()
-        self.private_registration = RegistrationFactory(project=self.private_project, user=self.user)
-        self.private_registration_wiki_id = self.private_registration.wiki_pages_versions['home'][0]
-        self.private_registration.wiki_pages_current = {'home': self.private_registration_wiki_id}
-        self.private_registration.save()
-        self.private_registration_url = '/{}wikis/{}/'.format(API_BASE, self.private_registration_wiki_id)
+        raise NotImplementedError
 
     def test_public_node_logged_out_user_can_view_wiki(self):
         self._set_up_public_project_with_wiki_page()
@@ -84,7 +62,8 @@ class TestWikiDetailView(ApiTestCase):
         private_link = PrivateLinkFactory(anonymous=True)
         private_link.nodes.append(self.private_project)
         private_link.save()
-        res = self.app.get('/{}wikis/{}/'.format(API_BASE, self.private_wiki._id), {'view_only': private_link.key})
+        url = furl.furl(self.private_url).add(query_params={'view_only': private_link.key}).url
+        res = self.app.get(url)
         assert_equal(res.status_code, 200)
         assert_equal(res.json['data']['id'], self.private_wiki._id)
 
@@ -93,7 +72,8 @@ class TestWikiDetailView(ApiTestCase):
         private_link = PrivateLinkFactory(anonymous=False)
         private_link.nodes.append(self.private_project)
         private_link.save()
-        res = self.app.get('/{}wikis/{}/'.format(API_BASE, self.private_wiki._id), {'view_only': private_link.key})
+        url = furl.furl(self.private_url).add(query_params={'view_only': private_link.key}).url
+        res = self.app.get(url)
         assert_equal(res.status_code, 200)
         assert_equal(res.json['data']['id'], self.private_wiki._id)
 
@@ -156,6 +136,35 @@ class TestWikiDetailView(ApiTestCase):
         expected_url = '/{}nodes/{}/comments/?filter[target]={}'.format(API_BASE, self.public_project._id, self.public_wiki._id)
         assert_equal(res.status_code, 200)
         assert_in(expected_url, url)
+
+
+class TestWikiDetailView(ApiWikiTestCase, TestWikiDetailMixin):
+
+    def _set_up_public_project_with_wiki_page(self):
+        self.public_project = ProjectFactory(is_public=True, creator=self.user)
+        self.public_wiki = self._add_project_wiki_page(self.public_project, self.user)
+        self.public_url = '/{}wikis/{}/'.format(API_BASE, self.public_wiki._id)
+
+    def _set_up_private_project_with_wiki_page(self):
+        self.private_project = ProjectFactory(creator=self.user)
+        self.private_wiki = self._add_project_wiki_page(self.private_project, self.user)
+        self.private_url = '/{}wikis/{}/'.format(API_BASE, self.private_wiki._id)
+
+    def _set_up_public_registration_with_wiki_page(self):
+        self._set_up_public_project_with_wiki_page()
+        self.public_registration = RegistrationFactory(project=self.public_project, user=self.user, is_public=True)
+        self.public_registration_wiki_id = self.public_registration.wiki_pages_versions['home'][0]
+        self.public_registration.wiki_pages_current = {'home': self.public_registration_wiki_id}
+        self.public_registration.save()
+        self.public_registration_url = '/{}wikis/{}/'.format(API_BASE, self.public_registration_wiki_id)
+
+    def _set_up_private_registration_with_wiki_page(self):
+        self._set_up_private_project_with_wiki_page()
+        self.private_registration = RegistrationFactory(project=self.private_project, user=self.user)
+        self.private_registration_wiki_id = self.private_registration.wiki_pages_versions['home'][0]
+        self.private_registration.wiki_pages_current = {'home': self.private_registration_wiki_id}
+        self.private_registration.save()
+        self.private_registration_url = '/{}wikis/{}/'.format(API_BASE, self.private_registration_wiki_id)
 
     def test_wiki_invalid_id_not_found(self):
         url = '/{}wikis/{}/'.format(API_BASE, 'abcde')
