+++ tests/base.py
@@ -335,6 +335,19 @@ class AdminTestCase(DbTestCase, DjangoTestCase, UploadTestCase, MockRequestTestC
     pass
 
 
+class ApiWikiTestCase(ApiTestCase):
+
+    def setUp(self):
+        from tests.factories import AuthUserFactory
+        super(ApiWikiTestCase, self).setUp()
+        self.user = AuthUserFactory()
+        self.non_contributor = AuthUserFactory()
+
+    def _add_project_wiki_page(self, node, user):
+        from tests.factories import NodeWikiFactory
+        # API will only return current wiki pages
+        return NodeWikiFactory(node=node, user=user, is_current=True)
+
 # From Flask-Security: https://github.com/mattupstate/flask-security/blob/develop/flask_security/utils.py
 class CaptureSignals(object):
     """Testing utility for capturing blinker signals.
